Added response codes as well as incrementing and decrementing respective counters for add and remove endpoints

+++ routes/removeCard.js
@@ -17,19 +17,45 @@ router.post('/api/removeCard', async (req, res, next) =>
     const { _id } = req.body;
 
     var error = "";
+    var set_id = [];
 
     try
     {
         const db = client.db();
+
+        // Find results
+        var tmp = await db.collection("Cards").find(
+            {"_id":mongo.ObjectID(_id)}
+        ).forEach(function(row)
+        {
+            console.log(row.set_id)
+            set_id.push(row.set_id);
+        })
+
+        // If no matching ID was found
+        if (set_id.length == 0)
+        {
+            res.status(400).json( {error:"This record doesn't exist. (Invalid ID)"})
+            return;
+        }
+
         const result = db.collection('Cards').deleteOne({_id:mongo.ObjectID(_id)}, function(err, result)
         {
             if (err != null)
                 error = err;
         });
+
+        // Remove one from num_sets
+        const updateNumber = db.collection('Sets').findOneAndUpdate(
+            { "_id":mongo.ObjectID(set_id[0])},
+            { $inc : { "num_cards" : -1 } }
+            );
     }
     catch(e)
     {
         error = e.toString();
+        res.status(500).json({ error:error });
+        return;
     }
 
     var ret = { error:error };
