+++ routes/removeSet.js
@@ -1,3 +1,4 @@
+const { json } = require('express');
 var express = require('express');
 var router = express.Router();
 var app = express();
@@ -17,19 +18,52 @@ router.post('/api/removeSet', async (req, res, next) =>
     const { _id } = req.body;
 
     var error = "";
+    var user_id = [];
+
+    if (_id == null)
+    {
+        error = "One or more needed fields are null. Check that your JSON Payload has the correct variables. (Requires: _id)";
+        res.status(400).json({ error:error });
+        return;
+    }
 
     try
     {
         const db = client.db();
+        
+        // Find results
+        var tmp = await db.collection("Sets").find(
+            {"_id":mongo.ObjectID(_id)}
+        ).forEach(function(row)
+        {
+            user_id.push(row.user_id);
+        })
+
+        // If no matching ID was found
+        if (user_id.length == 0)
+        {
+            res.status(400).json( {error:"This record doesn't exist. (Invalid ID)"})
+            return;
+        }
+
+        // Remove document
         const result = db.collection('Sets').deleteOne({_id:mongo.ObjectID(_id)}, function(err, result)
         {
             if (err != null)
                 error = err;
         });
+
+        // Remove one from num_sets
+        const updateNumber = db.collection('Users').findOneAndUpdate(
+            { "_id":user_id[0]},
+            { $inc : { "num_sets" : -1 } }
+            );
     }
     catch(e)
     {
         error = e.toString();
+        res.status(500).json( {error:error} );
+        return;
     }
 
     var ret = { error:error };
